Add the Tagging function to the Task objects
Now, a tag can be added to the different Task objects. For now, it is only
one, but a future extension can be the addition of multiple tags.

Tags make it easier to categorise certain similar tasks. Another
future extension to the existing Find functionality can be to
find tasks tagged with a particular tag.